Урок 4. Макеты Переименовали пакет viewPager в solarSystem
Попробовали параметры bias и percent
