+++ app/src/main/java/by/ealipatov/kotlin/materialyoufromealipatov/view/solarSystem/EarthFragment.kt
@@ -1,4 +1,4 @@
-package by.ealipatov.kotlin.materialyoufromealipatov.view.ViewPager
+package by.ealipatov.kotlin.materialyoufromealipatov.view.solarSystem
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import by.ealipatov.kotlin.materialyoufromealipatov.R
 
-class MarsFragment: Fragment() {
+class EarthFragment: Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_mars, container, false)
+        return inflater.inflate(R.layout.fragment_earth, container, false)
     }
 }
